Extend useEntryState hook to also return chapter data
* Refactor `sectionsWithNestedContentElements` to `entryStructure`
  which nests each section under their respective chapter parent.

* Move `useEntryState`-hook up to index for later use in `AppHeader`.
  Pass `entryStructure` and `dispatch` to `Entry` as props.

* Reflect changes in `Entry` by rendering newly introduced `Chapter`s
  with `anchor` props which can later be referred to in `AppHeader`
  instead of the `anchor`s currently attached to `Heading`s

* Move rendering of scenes and scene-specific logic from `Entry` to
  `Chapter`

REDMINE-17300
